api/requirements.txt
annotated-types>=0.6.0
fastapi>=0.115.0
pydantic>=2.11.0
pydantic-extra-types>=2.9.0
python-dotenv>=1.0.0
redis>=5.0.0
uvicorn[standard]>=0.30.0
